chore(util): move to separate directory
