+++ nuxt/components/form/FormInvitation.vue
@@ -83,10 +83,7 @@ import { useVuelidate } from '@vuelidate/core'
 import { minValue, required } from '@vuelidate/validators'
 import consola from 'consola'
 
-import { ITEMS_PER_PAGE_LARGE } from '~/plugins/util/constants'
-import { formPreSubmit } from '~/plugins/util/validation'
 import { Contact } from '~/types/contact'
-import { getApiMeta } from '~/plugins/util/util'
 import {
   useAllContactsQuery,
   useCreateInvitationMutation,
