+++ nuxt/components/invitation/InvitationList.vue
@@ -195,9 +195,7 @@ import consola from 'consola'
 import Swal from 'sweetalert2'
 import { Doughnut } from 'vue-chartjs/dist/index' // TODO: https://github.com/apertureless/vue-chartjs/pull/934
 
-import { copyText, getApiMeta } from '~/plugins/util/util'
 import { Invitation } from '~/types/invitation'
-import { ITEMS_PER_PAGE_LARGE } from '~/plugins/util/constants'
 import {
   useAllInvitationsQuery,
   useDeleteInvitationByIdMutation,
