+++ nuxt/composables/useHost.ts
@@ -1,5 +1,3 @@
-import { getHost } from '~/plugins/util/util'
-
 export const useHost = () => {
   if (process.server) {
     const event = useRequestEvent()
