+++ nuxt/cypress/e2e/taskAccountPasswordReset.cy.ts
@@ -1,4 +1,4 @@
-import { CYPRESS_BASE_URL } from '~/plugins/util/constants'
+import { CYPRESS_BASE_URL } from '~/utils/constants'
 
 describe('task account password reset page', () => {
   context('page load', () => {
