+++ nuxt/pages/dashboard/dev/index.vue
@@ -42,7 +42,6 @@
 </template>
 
 <script setup lang="ts">
-import { useSignOut } from '~/plugins/util/auth'
 import { useMaevsiStore } from '~/store'
 
 const { signOut } = useSignOut()
