+++ nuxt/utils/auth.ts
