+++ nuxt/utils/model.ts
